+++ CHANGELOG.md
@@ -9,3 +9,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Runtime upgraded to nodejs.12x
 - Base hook updated to the last configurations
 - New authorizers added
+- Event listener hook added